# openapi/api.yml
openapi: 3.0.0
info:
  title: GigaChill
  version: 0.1.0
servers:
  - url: https://giga-chill.ru/api
  - url: http://localhost:8081
security:
  - CookieAuth: []
...
paths:
...
  /events:
    get:
      summary: Список моих мероприятий
      description:
        Возвращает все мероприятия, в которых участвует аутентифицированный
        пользователь, вместе с его ролью в этих мероприятиях.
      tags: [Events]
      operationId: getEvents
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Events"
        "401":
          $ref: "#/components/responses/UnauthorizedError"

    post:
      summary: Создать мероприятие
      description: Автор запроса автоматически становится **owner**.
      tags: [Events]
      operationId: createEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventCreate"
      responses:
        "204":
          description: No Content
        "400":
          $ref: "#/components/responses/InvalidDataError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /events/{eventId}:
    get:
      summary: Получить мероприятие
      description:
        Возвращает мероприятие, в котором участвует аутентифицированный
        пользователь, вместе с его ролью в этом мероприятии.
      tags: [Events]
      operationId: getEvent
      parameters:
        - $ref: "#/components/parameters/eventId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

    patch:
      summary: Обновить мероприятие
      tags: [Events]
      operationId: updateEvent
      parameters:
        - $ref: "#/components/parameters/eventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "204":
          description: No Content
        "400":
          $ref: "#/components/responses/InvalidDataError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

    delete:
      summary: Удалить мероприятие
      tags: [Events]
      operationId: deleteEvent
      parameters:
        - $ref: "#/components/parameters/eventId"
      responses:
        "204":
          description: No Content
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /events/{eventId}/invitation-token:
    post:
      summary: Создать новый токен
      description: Токен генерирует пользователь со статусом **owner**.
      tags: [Invitation Tokens]
      operationId: createInvitationToken
      parameters:
        - $ref: "#/components/parameters/eventId"
      responses:
        "204":
          description: No Content
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

    get:
      summary: Получение токена
      description: Возвращает заранее сгенерированный токен-приглашение.
        Доступно пользователям с ролью **admin** или **owner**.
        Если токен ещё не был сгенерирован, то в поле **invitation_token** возвращается **null**.
      tags: [Invitation Tokens]
      operationId: getInvitationToken
      parameters:
        - $ref: "#/components/parameters/eventId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationToken"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /events/join-by-invitation-token:
    post:
      summary: Присоединиться к мероприятию по токену
      description: Пользователь становится участником мероприятия.
      tags: [Invitation Tokens]
      operationId: joinByInvitationToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvitationTokenJoin"
      responses:
        "200":
          description: Идентификатор мероприятия, к которому присоединился пользователь
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventId"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          description: Пользователь уже является участником мероприятия
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/{eventId}/participants:
    get:
      summary: Список участников
      tags: [Participants]
      operationId: getParticipants
      parameters:
        - $ref: "#/components/parameters/eventId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Participants"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

    post:
      summary: Добавить участника
      description: Добавление по логину. Роль по умолчанию — **participant**.
      tags: [Participants]
      operationId: addParticipant
      parameters:
        - $ref: "#/components/parameters/eventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParticipantCreate"
      responses:
        "204":
          description: No Content
        "400":
          $ref: "#/components/responses/InvalidDataError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          description: Пользователь уже является участником мероприятия
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/{eventId}/participants/{participantId}:
    delete:
      summary: Удалить участника
      tags: [Participants]
      operationId: deleteParticipant
      parameters:
        - $ref: "#/components/parameters/eventId"
        - $ref: "#/components/parameters/participantId"
      responses:
        "204":
          description: No Content
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /events/{eventId}/participants/{participantId}/role:
    patch:
      summary: Изменить роль участника
      tags: [Participants]
      operationId: setParticipantRole
      parameters:
        - $ref: "#/components/parameters/eventId"
        - $ref: "#/components/parameters/participantId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParticipantSetRole"
      responses:
        "204":
          description: No Content
        "400":
          $ref: "#/components/responses/InvalidDataError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
...
  /events/{eventId}/finalize:
    post:
      summary: Завершить мероприятие и провести расчёты долгов участников
      tags: [Debts]
      operationId: finalizeEvent
      parameters:
        - $ref: "#/components/parameters/eventId"
      responses:
        "204":
          description: No Content
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /events/{eventId}/my-balance:
    get:
      summary: Получить баланс пользователя в мероприятии
      tags: [Debts]
      operationId: getMyBalance
      parameters:
        - $ref: "#/components/parameters/eventId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserBalance"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /events/{eventId}/balance-summary:
    get:
      summary: Получить сводку по балансу мероприятия
      tags: [Debts]
      operationId: getBalanceSummary
      parameters:
        - $ref: "#/components/parameters/eventId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventBalanceSummary"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
...
components:
...
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: token
...